Calculate stint information
I don't think there is a way to do this within a single query. When it comes to
optimizations, this might be best as a view, or a stored procedure

Also, send pitStop information greedily within /api/statistics/entry/:entryId.
The additional endpoint is still going to be necessary to allow for comparisons